# contracts/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "commercial_v1_ConfirmSubscriptionChangeRequest": {
        "properties": {
          "changeId": {
            "type": "string"
          },
          "previewHash": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.ConfirmSubscriptionChangeRequest"
      },
...
      "commercial_v1_EntitlementCatalogVersion": {
        "properties": {
          "moduleCode": {
            "type": "string"
          },
          "version": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.EntitlementCatalogVersion"
      },
      "commercial_v1_EntitlementDecisionDTO": {
        "properties": {
          "allowed": {
            "type": "boolean"
          },
          "fieldAction": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "limit": {
            "$ref": "#/components/schemas/commercial_v1_EntitlementLimit"
          },
          "masked": {
            "type": "boolean"
          },
          "moduleCode": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/commercial_v1_EntitlementSourceExplanation"
            },
            "type": "array"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.EntitlementDecisionDTO"
      },
...
      "commercial_v1_EntitlementLimit": {
        "properties": {
          "unlimited": {
            "type": "boolean"
          },
          "value": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.EntitlementLimit"
      },
...
      "commercial_v1_EntitlementSourceExplanation": {
        "properties": {
          "actorId": {
            "type": "string"
          },
          "disposition": {
            "type": "string"
          },
          "effect": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "sourceKind": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.EntitlementSourceExplanation"
      },
...
      "commercial_v1_EntitlementView": {
        "properties": {
          "catalogRevision": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "catalogVersions": {
            "items": {
              "$ref": "#/components/schemas/commercial_v1_EntitlementCatalogVersion"
            },
            "type": "array"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/commercial_v1_EntitlementDecisionDTO"
            },
            "type": "array"
          },
          "entitlementVersion": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "evaluatedAt": {
            "type": "string"
          },
          "nextTransitionAt": {
            "type": "string"
          },
          "permissionSubject": {
            "type": "string"
          },
          "permissionVersion": {
            "type": "string"
          },
          "resolverVersion": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "sourceVersion": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "tenantId": {
            "type": "string"
          },
          "validUntil": {
            "type": "string"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.EntitlementView"
      },
...
      "commercial_v1_PlanField": {
        "properties": {
          "action": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.PlanField"
      },
      "commercial_v1_PlanModule": {
        "properties": {
          "capabilityCodes": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/commercial_v1_PlanField"
            },
            "type": "array"
          },
          "moduleCode": {
            "type": "string"
          },
          "quotas": {
            "items": {
              "$ref": "#/components/schemas/commercial_v1_PlanQuota"
            },
            "type": "array"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.PlanModule"
      },
      "commercial_v1_PlanQuota": {
        "properties": {
          "key": {
            "type": "string"
          },
          "unlimited": {
            "type": "boolean"
          },
          "value": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.PlanQuota"
      },
      "commercial_v1_PlanTerms": {
        "properties": {
          "modules": {
            "items": {
              "$ref": "#/components/schemas/commercial_v1_PlanModule"
            },
            "type": "array"
          },
          "priceRef": {
            "type": "string"
          },
          "salesScope": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "validityDays": {
            "minimum": 0,
            "type": "integer"
          },
          "validityMode": {
            "type": "string"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.PlanTerms"
      },
      "commercial_v1_PlanVersionDTO": {
        "properties": {
          "actorId": {
            "type": "string"
          },
          "contentSha256": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "planCode": {
            "type": "string"
          },
          "planRevision": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "publishedAt": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "retiredAt": {
            "type": "string"
          },
          "revision": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "state": {
            "type": "string"
          },
          "terms": {
            "$ref": "#/components/schemas/commercial_v1_PlanTerms"
          },
          "version": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.PlanVersionDTO"
      },
      "commercial_v1_PreviewSubscriptionChangeRequest": {
        "properties": {
          "action": {
            "type": "string"
          },
          "effectiveAt": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "targetPlanCode": {
            "type": "string"
          },
          "targetPlanVersion": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "tenantId": {
            "type": "string"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.PreviewSubscriptionChangeRequest"
      },
...
      "commercial_v1_ProvisioningRequirementDTO": {
        "properties": {
          "adapter": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "maxAttempts": {
            "minimum": 0,
            "type": "integer"
          },
          "version": {
            "type": "string"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.ProvisioningRequirementDTO"
      },
...
      "commercial_v1_SubscriptionChangeDependency": {
        "properties": {
          "moduleCode": {
            "type": "string"
          },
          "requiresModules": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.SubscriptionChangeDependency"
      },
      "commercial_v1_SubscriptionChangePreviewDTO": {
        "properties": {
          "action": {
            "type": "string"
          },
          "actorId": {
            "type": "string"
          },
          "before": {
            "$ref": "#/components/schemas/commercial_v1_TenantSubscriptionDTO"
          },
          "catalogRevision": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "changeId": {
            "type": "string"
          },
          "classification": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "currentEntitlements": {
            "$ref": "#/components/schemas/commercial_v1_EntitlementView"
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/commercial_v1_SubscriptionChangeDependency"
            },
            "type": "array"
          },
          "effectiveAt": {
            "type": "string"
          },
          "entitlementExpiresAt": {
            "type": "string"
          },
          "entitlementVersion": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "expiresAt": {
            "type": "string"
          },
          "impacts": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "mode": {
            "type": "string"
          },
          "previewHash": {
            "type": "string"
          },
          "pricingBasis": {
            "type": "string"
          },
          "projectedEntitlements": {
            "$ref": "#/components/schemas/commercial_v1_EntitlementView"
          },
          "provisioningRequirements": {
            "items": {
              "$ref": "#/components/schemas/commercial_v1_ProvisioningRequirementDTO"
            },
            "type": "array"
          },
          "quotaImpacts": {
            "items": {
              "$ref": "#/components/schemas/commercial_v1_SubscriptionChangeQuotaImpact"
            },
            "type": "array"
          },
          "quotaValidationRequired": {
            "type": "boolean"
          },
          "requestId": {
            "type": "string"
          },
          "sourceVersion": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "subscriptionRevision": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "target": {
            "$ref": "#/components/schemas/commercial_v1_PlanVersionDTO"
          },
          "tenantId": {
            "type": "string"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.SubscriptionChangePreviewDTO"
      },
      "commercial_v1_SubscriptionChangeQuotaImpact": {
        "properties": {
          "afterLimit": {
            "$ref": "#/components/schemas/commercial_v1_EntitlementLimit"
          },
          "beforeLimit": {
            "$ref": "#/components/schemas/commercial_v1_EntitlementLimit"
          },
          "evidence": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "moduleCode": {
            "type": "string"
          },
          "overLimit": {
            "type": "boolean"
          },
          "policy": {
            "type": "string"
          },
          "usageKnown": {
            "type": "boolean"
          },
          "used": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.SubscriptionChangeQuotaImpact"
      },
      "commercial_v1_SubscriptionChangeReceiptDTO": {
        "properties": {
          "action": {
            "type": "string"
          },
          "actorId": {
            "type": "string"
          },
          "after": {
            "$ref": "#/components/schemas/commercial_v1_TenantSubscriptionDTO"
          },
          "afterEntitlementVersion": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "afterSourceVersion": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "before": {
            "$ref": "#/components/schemas/commercial_v1_TenantSubscriptionDTO"
          },
          "beforeEntitlementVersion": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "beforeSourceVersion": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "changeId": {
            "type": "string"
          },
          "confirmedAt": {
            "type": "string"
          },
          "effectiveAt": {
            "type": "string"
          },
          "entitlementExpiresAt": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "previewHash": {
            "type": "string"
          },
          "pricingAuthority": {
            "type": "string"
          },
          "provisioningTaskId": {
            "type": "string"
          },
          "quotaImpacts": {
            "items": {
              "$ref": "#/components/schemas/commercial_v1_SubscriptionChangeQuotaImpact"
            },
            "type": "array"
          },
          "quotaValidationRequired": {
            "type": "boolean"
          },
          "reason": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.SubscriptionChangeReceiptDTO"
      },
      "commercial_v1_TenantSubscriptionDTO": {
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "entitlementSourceVersion": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "kind": {
            "type": "string"
          },
          "matchExplanation": {
            "type": "string"
          },
          "pendingChangeId": {
            "type": "string"
          },
          "periodEnd": {
            "type": "string"
          },
          "periodStart": {
            "type": "string"
          },
          "planCode": {
            "type": "string"
          },
          "planVersion": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "renewalStopped": {
            "type": "boolean"
          },
          "revision": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "ruleId": {
            "type": "string"
          },
          "ruleVersion": {
            "format": "uint64",
            "type": "string",
            "x-yunka-protobuf-json-integer": true
          },
          "salesScope": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "subscriptionId": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          }
        },
        "type": "object",
        "x-yunka-proto-full-name": "commercial.v1.TenantSubscriptionDTO"
      },
...
    }
  },
  "info": {
    "title": "yunka API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/platform/tenants/{tenant_id}/subscription/change-previews": {
      "post": {
        "operationId": "commercial.v1.SubscriptionChangesApplication.PreviewSubscriptionChange",
        "parameters": [
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/commercial_v1_PreviewSubscriptionChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/commercial_v1_SubscriptionChangePreviewDTO"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "tags": [
          "commercial.v1.SubscriptionChangesApplication"
        ],
        "x-yunka-rpc": "/commercial.v1.SubscriptionChangesApplication/PreviewSubscriptionChange"
      }
    },
    "/v1/platform/tenants/{tenant_id}/subscription/change-previews/{change_id}": {
      "get": {
        "operationId": "commercial.v1.SubscriptionChangesApplication.GetSubscriptionChangePreview",
        "parameters": [
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "changeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/commercial_v1_SubscriptionChangePreviewDTO"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "tags": [
          "commercial.v1.SubscriptionChangesApplication"
        ],
        "x-yunka-rpc": "/commercial.v1.SubscriptionChangesApplication/GetSubscriptionChangePreview"
      }
    },
    "/v1/platform/tenants/{tenant_id}/subscription/changes/{change_id}": {
      "get": {
        "operationId": "commercial.v1.SubscriptionChangesApplication.GetSubscriptionChangeReceipt",
        "parameters": [
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "changeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/commercial_v1_SubscriptionChangeReceiptDTO"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "tags": [
          "commercial.v1.SubscriptionChangesApplication"
        ],
        "x-yunka-rpc": "/commercial.v1.SubscriptionChangesApplication/GetSubscriptionChangeReceipt"
      }
    },
    "/v1/platform/tenants/{tenant_id}/subscription/changes/{change_id}/confirm": {
      "post": {
        "operationId": "commercial.v1.SubscriptionChangesApplication.ConfirmSubscriptionChange",
        "parameters": [
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "changeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/commercial_v1_ConfirmSubscriptionChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/commercial_v1_SubscriptionChangeReceiptDTO"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "tags": [
          "commercial.v1.SubscriptionChangesApplication"
        ],
        "x-yunka-rpc": "/commercial.v1.SubscriptionChangesApplication/ConfirmSubscriptionChange"
      }
    },
...
  },
...
}