# internal/commercial/transport/rest/zz_yunka_subscription_changes_operation_executor_gen.go
// Code generated by yunka contract; DO NOT EDIT.

package rest

import (
	"errors"
	commercialv1 "github.com/hvritual/biz/contracts/gen/commercial/v1"
	application "github.com/hvritual/biz/internal/commercial/application"
	policy "github.com/hvritual/biz/internal/commercial/policy"
	codes "google.golang.org/grpc/codes"
	status "google.golang.org/grpc/status"
	protojson "google.golang.org/protobuf/encoding/protojson"
	io "io"
	"net/http"
	execution "yunka.io/framework/execution"
	operation "yunka.io/framework/operation"
	authz "yunka.io/gateway/authz"
	httpbinding "yunka.io/gateway/httpbinding"
)

type SubscriptionChangesOperationHandler struct {
	application application.SubscriptionChangesApplication
	executor    operation.Executor
}

func RegisterSubscriptionChangesOperationExecutor(mux *http.ServeMux, application application.SubscriptionChangesApplication, executor operation.Executor) error {
	if mux == nil {
		return errors.New("contract C9 REST adapter: mux is required")
	}
	if application == nil {
		return errors.New("contract C9 REST adapter: application is required")
	}
	if executor == nil {
		return errors.New("contract C9 REST adapter: operation executor is required")
	}
	handler := &SubscriptionChangesOperationHandler{application: application, executor: executor}
	if err := httpbinding.Register(mux, "POST", "/v1/platform/tenants/{tenant_id}/subscription/changes/{change_id}/confirm", handler.handleOperationConfirmSubscriptionChange); err != nil {
		return err
	}
	if err := httpbinding.Register(mux, "GET", "/v1/platform/tenants/{tenant_id}/subscription/change-previews/{change_id}", handler.handleOperationGetSubscriptionChangePreview); err != nil {
		return err
	}
	if err := httpbinding.Register(mux, "GET", "/v1/platform/tenants/{tenant_id}/subscription/changes/{change_id}", handler.handleOperationGetSubscriptionChangeReceipt); err != nil {
		return err
	}
	if err := httpbinding.Register(mux, "POST", "/v1/platform/tenants/{tenant_id}/subscription/change-previews", handler.handleOperationPreviewSubscriptionChange); err != nil {
		return err
	}
	return nil
}

func writeSubscriptionChangesOperationError(writer http.ResponseWriter, err error) {
	if authz.IsDenied(err) {
		statusCode := http.StatusForbidden
		var denied *authz.DeniedError
		if errors.As(err, &denied) && (denied.Decision.Reason == authz.ReasonUnauthenticated || denied.Decision.Reason == authz.ReasonAuthenticationMethod) {
			statusCode = http.StatusUnauthorized
		}
		http.Error(writer, http.StatusText(statusCode), statusCode)
		return
	}
	if errors.Is(err, execution.ErrIdempotencyKeyRequired) {
		http.Error(writer, "idempotency key required", http.StatusBadRequest)
		return
	}
	if errors.Is(err, execution.ErrIdempotencyInProgress) || errors.Is(err, execution.ErrIdempotencyCompleted) {
		http.Error(writer, "idempotency conflict", http.StatusConflict)
		return
	}
	if code := status.Code(err); code == codes.Aborted || code == codes.AlreadyExists {
		http.Error(writer, "application conflict", http.StatusConflict)
		return
	}
	if errors.Is(err, operation.ErrExecutorUnavailable) || errors.Is(err, operation.ErrSecurityUnavailable) || errors.Is(err, operation.ErrSecurityNilContext) || errors.Is(err, operation.ErrIdempotencyUnavailable) {
		http.Error(writer, "operation execution unavailable", http.StatusInternalServerError)
		return
	}
	http.Error(writer, "application request failed", http.StatusBadRequest)
}

func (handler *SubscriptionChangesOperationHandler) handleOperationConfirmSubscriptionChange(writer http.ResponseWriter, request *http.Request) {
	wire := &commercialv1.ConfirmSubscriptionChangeRequest{}
	body, err := io.ReadAll(request.Body)
	if err != nil {
		http.Error(writer, "invalid request body", http.StatusBadRequest)
		return
	}
	if len(body) > 0 {
		if err := protojson.Unmarshal(body, wire); err != nil {
			http.Error(writer, "invalid request body", http.StatusBadRequest)
			return
		}
	}
	wire.TenantId = request.PathValue("tenant_id")
	wire.ChangeId = request.PathValue("change_id")
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanSubscriptionChangesConfirmSubscriptionChange(), wire, handler.application.ConfirmSubscriptionChange)
	if err != nil {
		writeSubscriptionChangesOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}

func (handler *SubscriptionChangesOperationHandler) handleOperationGetSubscriptionChangePreview(writer http.ResponseWriter, request *http.Request) {
	wire := &commercialv1.ReadSubscriptionChangeRequest{}
	wire.TenantId = request.PathValue("tenant_id")
	wire.ChangeId = request.PathValue("change_id")
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanSubscriptionChangesGetSubscriptionChangePreview(), wire, handler.application.GetSubscriptionChangePreview)
	if err != nil {
		writeSubscriptionChangesOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}

func (handler *SubscriptionChangesOperationHandler) handleOperationGetSubscriptionChangeReceipt(writer http.ResponseWriter, request *http.Request) {
	wire := &commercialv1.ReadSubscriptionChangeRequest{}
	wire.TenantId = request.PathValue("tenant_id")
	wire.ChangeId = request.PathValue("change_id")
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanSubscriptionChangesGetSubscriptionChangeReceipt(), wire, handler.application.GetSubscriptionChangeReceipt)
	if err != nil {
		writeSubscriptionChangesOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}

func (handler *SubscriptionChangesOperationHandler) handleOperationPreviewSubscriptionChange(writer http.ResponseWriter, request *http.Request) {
	wire := &commercialv1.PreviewSubscriptionChangeRequest{}
	body, err := io.ReadAll(request.Body)
	if err != nil {
		http.Error(writer, "invalid request body", http.StatusBadRequest)
		return
	}
	if len(body) > 0 {
		if err := protojson.Unmarshal(body, wire); err != nil {
			http.Error(writer, "invalid request body", http.StatusBadRequest)
			return
		}
	}
	wire.TenantId = request.PathValue("tenant_id")
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanSubscriptionChangesPreviewSubscriptionChange(), wire, handler.application.PreviewSubscriptionChange)
	if err != nil {
		writeSubscriptionChangesOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}